move types to dedicated package
