+++ contracts/ics721-base-tester/Cargo.toml
@@ -19,5 +19,6 @@ cosmwasm-storage = { workspace = true }
 cw-storage-plus = { workspace = true }
 cw2 = { workspace = true }
 ics721 = { workspace = true }
+ics721-types = { workspace = true }
 thiserror = { workspace = true }
 cw721 = { workspace = true }
