+++ contracts/sg-ics721/Cargo.toml
@@ -17,6 +17,7 @@ cosmwasm-std = { workspace = true, features = ["ibc3"] }
 cosmwasm-schema = { workspace = true }
 cw2 = { workspace = true }
 ics721 = { workspace = true }
+ics721-types = { workspace = true }
 sg-std = { workspace = true}
 sg721 = { workspace = true }
 sg721-base = { workspace = true, features = ["library"] }
