+++ packages/ics721-types/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "ics721-types"
+version = "0.1.0"
+edition = "2021"
+authors = ["ekez <ekez@withoutdoing.com>", "mr-t <mr-t@arkprotocol.io>", "Art3mix <art3mix.90@gmail.com>", "m@publicawesome.com", "s+git@publicawesome.com"]
+description = "an implementation of the ICS721 specification for transfering NFTs between blockchains"
+
+[dependencies]
+cosmwasm-std = { workspace = true, features = ["cosmwasm_1_2"] }
+cosmwasm-schema = { workspace = true }
+cw-storage-plus = { workspace = true }
+serde = { workspace = true }
+thiserror = { workspace = true }
