+++ packages/ics721-types/src/types.rs
