+++ packages/ics721/Cargo.toml
@@ -15,6 +15,7 @@ cw-utils = { workspace = true }
 cw721 = { workspace = true }
 cw721-base = { workspace = true, features = ["library"] }
 cw721-base-016  = { workspace = true, features = ["library"] }
+ics721-types = { workspace = true }
 thiserror = { workspace = true }
 serde = { workspace = true }
 cw-paginate-storage = { workspace = true }
