+++ packages/ics721/src/execute.rs
@@ -4,13 +4,14 @@ use cosmwasm_std::{
     from_json, to_json_binary, Addr, Binary, ContractInfoResponse, Deps, DepsMut, Empty, Env,
     IbcMsg, MessageInfo, Response, StdResult, SubMsg, WasmMsg,
 };
+use ics721_types::{token_types::{TokenId, Class, ClassId, Token}, ibc::NonFungibleTokenPacketData};
 use serde::{de::DeserializeOwned, Serialize};
 use sha2::{Digest, Sha256};
 
 use crate::{
     helpers::get_instantiate2_address,
     ibc::{
-        NonFungibleTokenPacketData, INSTANTIATE_CW721_REPLY_ID,
+        INSTANTIATE_CW721_REPLY_ID,
         INSTANTIATE_INCOMING_PROXY_REPLY_ID, INSTANTIATE_OUTGOING_PROXY_REPLY_ID,
     },
     msg::{CallbackMsg, ExecuteMsg, IbcOutgoingMsg, InstantiateMsg, MigrateMsg},
@@ -19,7 +20,7 @@ use crate::{
         CW721_CODE_ID, INCOMING_PROXY, NFT_CONTRACT_TO_CLASS_ID, OUTGOING_CLASS_TOKEN_TO_CHANNEL,
         OUTGOING_PROXY, PO, TOKEN_METADATA,
     },
-    token_types::{Class, ClassId, Token, TokenId, VoucherCreation, VoucherRedemption},
+    token_types::{VoucherCreation, VoucherRedemption},
     ContractError,
 };
