+++ packages/ics721/src/ibc.rs
@@ -1,11 +1,11 @@
-use cosmwasm_schema::cw_serde;
 use cosmwasm_std::{
-    from_json, to_json_binary, Binary, DepsMut, Empty, Env, IbcBasicResponse, IbcChannelCloseMsg,
+    from_json, to_json_binary, DepsMut, Empty, Env, IbcBasicResponse, IbcChannelCloseMsg,
     IbcChannelConnectMsg, IbcChannelOpenMsg, IbcChannelOpenResponse, IbcPacket, IbcPacketAckMsg,
     IbcPacketReceiveMsg, IbcPacketTimeoutMsg, IbcReceiveResponse, Never, Reply, Response,
     StdResult, SubMsgResult, WasmMsg,
 };
 use cw_utils::parse_reply_instantiate_data;
+use ics721_types::{ibc::NonFungibleTokenPacketData, types::Ics721Status};
 use serde::{de::DeserializeOwned, Serialize};
 
 use crate::{
@@ -16,8 +16,6 @@ use crate::{
         CLASS_ID_TO_NFT_CONTRACT, INCOMING_CLASS_TOKEN_TO_CHANNEL, INCOMING_PROXY,
         NFT_CONTRACT_TO_CLASS_ID, OUTGOING_CLASS_TOKEN_TO_CHANNEL, OUTGOING_PROXY, TOKEN_METADATA,
     },
-    token_types::{ClassId, TokenId},
-    types::Ics721Status,
     ContractError,
 };
 
@@ -35,41 +33,6 @@ pub(crate) const ACK_AND_DO_NOTHING: u64 = 2;
 pub(crate) const ACK_CALLBACK_REPLY_ID: u64 = 3;
 /// The IBC version this contract expects to communicate with.
 pub const IBC_VERSION: &str = "ics721-1";
-
-#[cw_serde]
-#[serde(rename_all = "camelCase")]
-pub struct NonFungibleTokenPacketData {
-    /// Uniquely identifies the collection which the tokens being
-    /// transfered belong to on the sending chain. Must be non-empty.
-    pub class_id: ClassId,
-    /// Optional URL that points to metadata about the
-    /// collection. Must be non-empty if provided.
-    pub class_uri: Option<String>,
-    /// Optional base64 encoded field which contains on-chain metadata
-    /// about the NFT class. Must be non-empty if provided.
-    pub class_data: Option<Binary>,
-    /// Uniquely identifies the tokens in the NFT collection being
-    /// transfered. This MUST be non-empty.
-    pub token_ids: Vec<TokenId>,
-    /// Optional URL that points to metadata for each token being
-    /// transfered. `tokenUris[N]` should hold the metadata for
-    /// `tokenIds[N]` and both lists should have the same if
-    /// provided. Must be non-empty if provided.
-    pub token_uris: Option<Vec<String>>,
-    /// Optional base64 encoded metadata for the tokens being
-    /// transfered. `tokenData[N]` should hold metadata for
-    /// `tokenIds[N]` and both lists should have the same length if
-    /// provided. Must be non-empty if provided.
-    pub token_data: Option<Vec<Binary>>,
-
-    /// The address sending the tokens on the sending chain.
-    pub sender: String,
-    /// The address that should receive the tokens on the receiving
-    /// chain.
-    pub receiver: String,
-    /// Memo to add custom string to the msg
-    pub memo: Option<String>,
-}
 
 pub trait Ics721Ibc<T = Empty>
 where
