+++ packages/ics721/src/ibc_helpers.rs
@@ -1,10 +1,11 @@
 use cosmwasm_std::{
     from_json, to_json_binary, Binary, IbcAcknowledgement, IbcChannel, IbcEndpoint, IbcOrder,
 };
+use ics721_types::ibc::NonFungibleTokenPacketData;
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    ibc::{NonFungibleTokenPacketData, IBC_VERSION},
+    ibc::IBC_VERSION,
     ContractError,
 };
 
@@ -120,54 +121,10 @@ pub(crate) fn validate_order_and_version(
     Ok(())
 }
 
-macro_rules! non_empty_optional {
-    ($e:expr) => {
-        if $e.map_or(false, |data| data.is_empty()) {
-            return Err(ContractError::EmptyOptional {});
-        }
-    };
-}
-
-impl NonFungibleTokenPacketData {
-    pub fn validate(&self) -> Result<(), ContractError> {
-        if self.class_id.is_empty() {
-            return Err(ContractError::EmptyClassId {});
-        }
-
-        non_empty_optional!(self.class_uri.as_ref());
-        non_empty_optional!(self.class_data.as_ref());
-
-        let token_count = self.token_ids.len();
-        if token_count == 0 {
-            return Err(ContractError::NoTokens {});
-        }
-
-        // Non-empty optionality of tokenData an tokenUris implicitly
-        // checked here.
-        if self
-            .token_data
-            .as_ref()
-            .map_or(false, |data| data.len() != token_count)
-            || self
-                .token_uris
-                .as_ref()
-                .map_or(false, |data| data.len() != token_count)
-        {
-            return Err(ContractError::TokenInfoLenMissmatch {});
-        }
-
-        // This contract assumes that the backing cw721 is functional,
-        // so no need to check tokenIds for duplicates as the cw721
-        // will prevent minting of duplicates.
-
-        Ok(())
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::token_types::{ClassId, TokenId};
+    use ics721_types::token_types::{ClassId, TokenId};
 
     #[test]
     fn test_pop_source_simple() {
@@ -220,76 +177,5 @@ mod tests {
             ),
             None
         );
-    }
-
-    #[test]
-    fn test_packet_validation() {
-        let default_token = NonFungibleTokenPacketData {
-            class_id: ClassId::new("id"),
-            class_uri: None,
-            class_data: None,
-            token_ids: vec![TokenId::new("1")],
-            token_uris: None,
-            token_data: None,
-            sender: "violet".to_string(),
-            receiver: "blue".to_string(),
-            memo: None,
-        };
-
-        let empty_class_id = NonFungibleTokenPacketData {
-            class_id: ClassId::new(""),
-            ..default_token.clone()
-        };
-        let err = empty_class_id.validate().unwrap_err();
-        assert_eq!(err, ContractError::EmptyClassId {});
-
-        let empty_class_uri = NonFungibleTokenPacketData {
-            class_uri: Some("".to_string()),
-            ..default_token.clone()
-        };
-        let err = empty_class_uri.validate().unwrap_err();
-        assert_eq!(err, ContractError::EmptyOptional {});
-
-        let empty_class_data = NonFungibleTokenPacketData {
-            class_data: Some(Binary::default()),
-            ..default_token.clone()
-        };
-        let err = empty_class_data.validate().unwrap_err();
-        assert_eq!(err, ContractError::EmptyOptional {});
-
-        let no_tokens = NonFungibleTokenPacketData {
-            token_ids: vec![],
-            ..default_token.clone()
-        };
-        let err = no_tokens.validate().unwrap_err();
-        assert_eq!(err, ContractError::NoTokens {});
-
-        let uri_imbalance_empty = NonFungibleTokenPacketData {
-            token_uris: Some(vec![]),
-            ..default_token.clone()
-        };
-        let err = uri_imbalance_empty.validate().unwrap_err();
-        assert_eq!(err, ContractError::TokenInfoLenMissmatch {});
-
-        let uri_imbalance = NonFungibleTokenPacketData {
-            token_uris: Some(vec!["a".to_string(), "b".to_string()]),
-            ..default_token.clone()
-        };
-        let err = uri_imbalance.validate().unwrap_err();
-        assert_eq!(err, ContractError::TokenInfoLenMissmatch {});
-
-        let data_imbalance_empty = NonFungibleTokenPacketData {
-            token_data: Some(vec![]),
-            ..default_token.clone()
-        };
-        let err = data_imbalance_empty.validate().unwrap_err();
-        assert_eq!(err, ContractError::TokenInfoLenMissmatch {});
-
-        let data_imbalance = NonFungibleTokenPacketData {
-            token_data: Some(vec![Binary::default(), Binary::default()]),
-            ..default_token
-        };
-        let err = data_imbalance.validate().unwrap_err();
-        assert_eq!(err, ContractError::TokenInfoLenMissmatch {});
     }
 }
