+++ packages/ics721/src/lib.rs
@@ -8,9 +8,8 @@ pub mod msg;
 pub mod query;
 pub mod state;
 pub mod token_types;
-pub mod types;
 pub mod utils;
-
+use ics721_types::{token_types::{Class, ClassId, Token, TokenId}, ibc::NonFungibleTokenPacketData};
 pub use crate::error::ContractError;
 
 #[cfg(test)]
