+++ packages/ics721/src/testing/ibc_tests.rs
@@ -11,18 +11,17 @@ use cosmwasm_std::{
 use crate::{
     execute::Ics721Execute,
     ibc::{
-        Ics721Ibc, NonFungibleTokenPacketData, ACK_AND_DO_NOTHING, IBC_VERSION,
+        Ics721Ibc, ACK_AND_DO_NOTHING, IBC_VERSION,
         INSTANTIATE_CW721_REPLY_ID,
     },
     ibc_helpers::{ack_fail, ack_success, try_get_ack_error},
     msg::{CallbackMsg, ExecuteMsg, InstantiateMsg, QueryMsg},
     query::Ics721Query,
     state::{CollectionData, INCOMING_CLASS_TOKEN_TO_CHANNEL, NFT_CONTRACT_TO_CLASS_ID, PO},
-    token_types::{ClassId, TokenId},
-    types::Ics721Callbacks,
     utils::get_collection_data,
     ContractError,
 };
+use ics721_types::{token_types::{ClassId, TokenId}, ibc::NonFungibleTokenPacketData, types::Ics721Callbacks};
 
 const CONTRACT_PORT: &str = "wasm.address1";
 const REMOTE_PORT: &str = "stars.address1";
@@ -497,7 +496,7 @@ fn test_ibc_packet_receive_invalid_packet_data() {
 
     assert!(error
         .unwrap()
-        .starts_with("Error parsing into type ics721::ibc::NonFungibleTokenPacketData"))
+        .starts_with("Error parsing into type ics721_types::ibc::NonFungibleTokenPacketData"))
 }
 
 #[test]
@@ -558,7 +557,7 @@ fn test_ibc_packet_receive_missmatched_lengths() {
 
     assert_eq!(
         error,
-        Some(ContractError::TokenInfoLenMissmatch {}.to_string())
+        Some(ContractError::ValidationError(ics721_types::error::ValidationError::TokenInfoLenMissmatch{}).to_string())
     );
 
     // More token data are provided than tokens.
@@ -590,7 +589,7 @@ fn test_ibc_packet_receive_missmatched_lengths() {
 
     assert_eq!(
         error,
-        Some(ContractError::TokenInfoLenMissmatch {}.to_string())
+        Some(ContractError::ValidationError(ics721_types::error::ValidationError::TokenInfoLenMissmatch{}).to_string())
     )
 }
